decNumber.c (decStrEq): Cast operands to int before calling tolower to avoid compilation warnings on Tru64.
	* decNumber.c (decStrEq): Cast operands to int before calling
	tolower to avoid compilation warnings on Tru64.

From-SVN: r108858

+++ libdecnumber/decNumber.c
@@ -5445,7 +5445,7 @@ decStrEq (const char *str1, const char *str2)
 	}
       else
 	{
-	  if (tolower (*str1) != tolower (*str2))
+	  if (tolower ((int) *str1) != tolower ((int) *str2))
 	    return 0;
 	}
     }				/* stepping */
